requirements.txt
tiktoken
langchain==0.3.27
langchain-community
langchain-core
langchain-text-splitters
langchain-mistralai
chromadb
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
GitPython
python-dotenv
jupyter
chainlit